==> R/00_all.qmd <==
---
title: "00_all"
format: 
  html: 
    self-contained: true  # ingen _files-mapper
---

Please write quarto render R/00_all.qmd in the terminal to render all .qmd documents and generate a html that contains the full project. You can find the html file in the "results" folder.

# 01_load

```{r, child="01_load.qmd"}
```

# 02_clean

```{r, child="02_clean.qmd"}
```

# 03_augment

```{r, child="03_augment.qmd"}
```

# 04_describe

```{r, child="04_describe.qmd"}
```

# 05_analysis_1

```{r, child="05_analysis_1.qmd"}
```

# 06_analysis_2

```{r, child="06_analysis_2.qmd"}
```

# 07_analysis_3

```{r, child="07_analysis_3.qmd"}
```

# 08_analysis_4

```{r, child="08_analysis_4.qmd"}
```
