==> R/01_load.qmd <==
---
title: "01_load"
format: html
execute:
  standalone: true
---

```{r}
#| message: false

library("tidyverse")
```

```{r}
#| message: false
#| warning: false

raw_data <- read_csv("../data/_raw/SAA_Biospecimen_Analysis_Results_01Nov2025.csv")
```

```{r}
write_tsv(raw_data, "../data/01_dat_load.tsv")
```
