==> R/03_augment.qmd <==
---
title: "03_augment"
format: html
---

```{r}
#| message: false

library(tidyverse)
```

## Loading the clean data

```{r}
#| message: false

tidy_data <- read_tsv("../data/02_dat_clean.tsv")
```

## Augment Data

### Finding days from baseline

Some patients were tested multiple times at different dates. We calculate the days from baseline and add a colum with this information.

It should be noted that for 2 patients (40771 and 40754) the rundate appears to have been inputted wrong, so that it appears as if several samples were taken on just one day, despite them having different clinical_events. Therefore, it is insufficient to just sort by days_from_baseline, and clinical_events is therefore kept.

```{r}
tidy_data <- tidy_data |> 
  mutate(rundate = as.Date(rundate)) |> 
  group_by(patient_number) |> 
  mutate(days_from_baseline = as.numeric(rundate - min(rundate, 
                                                       na.rm = TRUE)
                                         )) |> 
  ungroup()
```

### Calculating means of Fmax, ttt and auc

Instead of working with three replicates of a single observation, moving forward the mean of the variable will be used to represent the observation.

The means are calculated of Fmax, ttt and auc from a single patient (patient_number) per visit (clinical_event).

```{r}
means <- tidy_data |>
  group_by(patient_number, clinical_event, days_from_baseline) |>
  
  summarise(
    fmax_mean = mean(fmax, na.rm = TRUE),
    ttt_mean  = mean(ttt, na.rm = TRUE),
    auc_mean  = mean(auc, na.rm = TRUE),
    .groups = "drop"
  )

aug_data <- tidy_data |>
  left_join(means,
            by = c("patient_number", 
                   "clinical_event", 
                   "days_from_baseline"))

```

### Initiating a quality check

When PPMI performed their study, they choose a criteria the Fmax value must meet, for the SAA-result to indicate PD. To ensure they correctly assigned positive/negative/inconclusive, the saa_results to the observations, a quality check is made.

As the two projects differ in method, they used different requirements for each.

|              |                                            |
|--------------|--------------------------------------------|
| **Result**   | **Project 155 - Criteria of Fmax**         |
| Positive     | All 3 replicates have Fmax ≥ 5,000 RFU     |
| Negative     | 0 or 1 replicate has Fmax ≥ 5,000 RFU      |
| Inconclusive | Exactly 2 replicates have Fmax ≥ 5,000 RFU |

+--------------+--------------------------------------------------------------------------------------------------------------------------------------+
| Result       | **Project 237 - Criteria of Fmax**                                                                                                   |
+==============+======================================================================================================================================+
| Positive     | -   All 3 replicates have Fmax ≥ 45,000 RFU                                                                                          |
|              |                                                                                                                                      |
|              |     or                                                                                                                               |
|              |                                                                                                                                      |
|              | -   Exactly 2 or 3 replicates have Fmax ≥ 3,000 RFU & \< 45,000 RFU                                                                  |
|              |                                                                                                                                      |
|              |     or                                                                                                                               |
|              |                                                                                                                                      |
|              | -   Exactly 2 replicates have Fmax ≥ 45,000 RFU & 1 replicate has Fmax ≥ 3,000 RFU & \< 45,000 RFU                                   |
+--------------+--------------------------------------------------------------------------------------------------------------------------------------+
| Negative     | -   0 or 1 replicate has Fmax ≥ 3,000 RFU                                                                                            |
+--------------+--------------------------------------------------------------------------------------------------------------------------------------+
| Inconclusive | -   Exactly 2 replicates have Fmax ≥ 45,000 RFU & 1 replicate has Fmax \< 3,000 RFU                                                  |
|              |                                                                                                                                      |
|              |     or                                                                                                                               |
|              |                                                                                                                                      |
|              | -   Exactly 1 replicate has Fmax ≥ 45,000 RFU & 1 replicate has Fmax ≥ 3,000 RFU & \< 45,000 RFU & 1 replicate has Fmax \< 3,000 RFU |
+--------------+--------------------------------------------------------------------------------------------------------------------------------------+

To prepare replicate vectors for SAA rule all three fmax from each observation is gathered in a list.

```{r}
source("09_proj_func.R")

reps <- tidy_data |>
  group_by(patient_number, 
           project, 
           clinical_event, 
           days_from_baseline) |> 
  
  summarise(
    fmax_reps = list(fmax),
    .groups = "drop"
  )

aug_data <- aug_data |> 
  left_join(reps,
            by = c("patient_number", 
                   "project", 
                   "clinical_event",
                   "days_from_baseline"))
```

The two rules have been made as functions,

Applying the functions and subsequently removing the generated fmax_reps column:

```{r}
source("09_proj_func.R")

aug_data <- aug_data |> 
  rowwise() |> 
  
  mutate(
    saa_custom = case_when(
      project == "155" ~ saa_rule_155(fmax_reps),
      project == "237" ~ saa_rule_237(fmax_reps),
      TRUE ~ NA_character_
    )
  ) |> 
  
  ungroup()

aug_data <- aug_data |>  
  select(-fmax_reps)

```

### Save the augmented data

```{r}
write_tsv(aug_data, "../data/03_aug_data.tsv")
```
