==> R/04_describe.qmd <==
---
title: "04_describe"
format: html
---

```{r}
#| message: false

library(tidyverse)
```

## Loading the augmented data

```{r}
#| message: false

aug_data <- read_tsv("../data/03_aug_data.tsv")
```

## Describing the data

To ensure uniformity across the different plots a general theme has been made:

```{r}
  theme_custom <- theme_minimal(base_size = 14) +
  theme(
    plot.title = element_text(face = "bold", size = 13, margin = margin(b = 7)),
    plot.subtitle = element_text(size = 10, margin = margin(b = 12)),
    plot.caption = element_text(size = 8, margin = margin(t = 20)),
      
    axis.title.x = element_text(size = 10, margin = margin(t = 10)),
    axis.title.y = element_text(size = 10, margin = margin(r = 10)),

    axis.text.x = element_text(size = 10, margin = margin(t = 5)),
    axis.text.y = element_text(size = 10, margin = margin(r = 5)),
    
    legend.position = "none",
    )
```

### Distribution of observations between projects

The data contains observations made in two projects. The distribution from each projects can be seen in the plot as well as the male/female ratio.

```{r}
project_distribution <- aug_data |>
  mutate(project = as.character(project)) |>
  ggplot(mapping = aes(x = project)) + 
  geom_bar(mapping = aes(fill = sex)) + 
  scale_fill_manual(values = c("#da9283","#3C3C68"),
                    name = "Sex") +
  theme_custom + 
  theme(legend.position = "bottom") + 
  labs(title = "Project Distribution",
       subtitle = "The distribution of observations and gender across project 155 and 237",
       x = "Project",
       y = "Observations",
       caption = "Source: PPMI (Parkinson's Progression Markers Initiative)")

ggsave(filename = "../results/04_project_distribution.png",
       width = 10,  
       height = 6)

project_distribution
```

### Individual participants

The data-set orignially contained more than 2000 observations. Some of the patients appear multiple times in the data set. The amount of different patients enrolled in the experiment is determined. Obs: we removed one patient during the cleaning because of a missing value interpretation.

```{r}
patients_participants <- aug_data |>
  distinct(patient_number) |>
  summarize(Participants = n())
patients_participants

```

### Number of observations

1 observation is rep 1-3

```{r}
observations <- aug_data |>
  group_by(clinical_event) |>
  distinct(patient_number) |>
  ungroup()|>
  summarize(Observations = n())
observations
```

### Result of the SAA

All patients enrolled was diagnosed with PD previously to SAA analysis. The result of the SAA does not follow the PD diagnosis completely. The number of saa-positive patients at their latest visit.

```{r}
aug_data |>
  arrange(desc(days_from_baseline)) |>
  distinct(patient_number, .keep_all = TRUE) |>
  filter(saa_result == "Positive") |>
  summarize(Positive_individuals = n())
  
```

### Saa-result from all observations, patient repeats accepted

```{r}
result_distribution <- aug_data |>
  mutate(project = as.character(project)) |>
  ggplot(mapping = aes(x = project)) + 
  geom_bar(mapping = aes(fill = saa_result)) + 
  scale_fill_manual(values = c("#7A1F15","#da9283","#94BFBE"),
                    name = "Result",
                    guide = guide_legend(nrow = 1)) +
  theme_custom + 
  theme(legend.position = "bottom") + 
  labs(title = "Result Distribution",
       subtitle = "The distribution of saa-result",
       x = "Project",
       y = "Observations",
       caption = "Source: PPMI (Parkinson's Progression Markers Initiative)")

ggsave(filename = "../results/04_result_distribution.png",
       width = 10,  
       height = 5)

result_distribution
```

### Patient repeats

The distribution of each patient's latest experiment was plotted. Most of the data only contains information from the Baseline experiment. If the patients went for further experiments, most of them got the last assessment \~12 months (V04) after their Baseline.

```{r}
multiple_experiments <- aug_data |>
  arrange(desc(days_from_baseline)) |>
  distinct(patient_number, .keep_all = TRUE) |>
  ggplot(mapping = aes(x = clinical_event,
                       fill = clinical_event)) + 
  geom_bar(show.legend = FALSE) + 
  scale_fill_manual(values = c("#3C3C68","#627634","#545E75","#C03221", "#94BFBE","#7A1F15","#da9283","#474B24","#474B24","#474B24","#474B24")) +
  theme_custom + 
  labs(title = "Latest observation",
       subtitle = "The latest observation of each patient",
       x = "Latest Observation",
       y = "Participants",
       caption = "Source: PPMI (Parkinson's Progression Markers Initiative)")

ggsave(filename = "../results/04_multiple_experiments.png",
       width = 10,  
       height = 5)

multiple_experiments
```

### Use of instrument

To analyse the samples of cerebrospinal fluid in the SAA, different equipment was used. In project 155 instrument 2, 4 and 6 were equally used. During project 237 these instruments are also used frequently in addition to instrument 5.

```{r}
instruments_used <- aug_data |>
  mutate(project = as.character(project)) |>
  mutate(instrument = as.character(instrument)) |>
  ggplot(mapping = aes(x = project)) + 
  geom_bar(mapping = aes(fill = instrument),
           position = "fill") + 
  scale_fill_manual(breaks = c("2","4","5","6","9","10"),
                    values = c("#da9283","#7A1F15","#C03221","#94BFBE","#3C3C68","#627634"),
                    name = "Instrument",
                    guide = guide_legend(nrow = 1)) +
  theme_custom + 
  theme(legend.position = "bottom") + 
  labs(title = "Used Instrument",
       subtitle = "The instrument used to analyze each observation",
       x = "Project",
       y = "Ratio",
       caption = "Source: PPMI (Parkinson's Progression Markers Initiative)")

ggsave(filename = "../results/04_instruments_used.png",
       width = 10,  
       height = 6)

instruments_used

```

### High and low ranges of key values

To get a better understanding of the ranges of key values (fmax, auc, ttt) we look at a few examples at each end of the spectrum.

```{r}
source("09_proj_func.R")

# The function returns 2 values: one from project 155 and one from 237:
low_range(aug_data, fmax, 50)
low_range(aug_data, auc, 50)
low_range(aug_data, ttt, 50)
```

```{r}
source("09_proj_func.R")

# The function returns 2 values: one from project 155 and one from 237:
high_range(aug_data, fmax, 50)
high_range(aug_data, auc, 50)
high_range(aug_data, ttt, 50)
```

```{r}
#| fig.height: 15
#| fig.width: 15

library(patchwork)

((instruments_used + project_distribution) / multiple_experiments / result_distribution) +
  plot_layout(guides = "collect",
              heights = c(5,5,7,5)) &
  theme(legend.position = "bottom")

```
